Fix GC race condition resulting in occasional empty adiff files

+++ merge.mk
@@ -19,9 +19,16 @@ changesets: $(shell find stage-data/split-adiffs/ -mindepth 1 -type d -printf 's
 stage-data/changesets/%.adiff.md5: $$(wildcard stage-data/split-adiffs/%/*)
 	tmpfile=$$(mktemp)
 	merge_adiffs.py $^ | xmlstarlet format > $$tmpfile
-	md5sum < $$tmpfile > $@
-	gzip -c < $$tmpfile > $$tmpfile.gz
-	mv $$tmpfile.gz bucket-data/changesets/$*.adiff && rm $$tmpfile
+	if [ -s $$tmpfile ]; then
+		# merge_adiffs.py can fail if it is given no input files or if one or more
+		# of its input files are not found. Either of these can happen if the input
+		# split-adiffs/*/ directory is deleted by gc.sh while Make is running this
+		# script. So we only move the adiff file into bucket-data and update the
+		# stamp file if the merged output file is nonempty (-s).
+		md5sum < $$tmpfile > $@
+		gzip -c < $$tmpfile > $$tmpfile.gz
+		mv $$tmpfile.gz bucket-data/changesets/$*.adiff && rm $$tmpfile
+	fi
 
 metadatas: $(shell find stage-data/split-adiffs/ -mindepth 1 -type d -printf '%p/metadata.xml\n')
